+++ botan-low/CHANGELOG.md
@@ -17,6 +17,9 @@
   experimental `x509` code is moved.
 * PATCH: enable `-Wall` in addition to a number of other GHC warnings.
 * PATCH: use `GHC2021` as the default language.
+* BREAKING: `Botan.Low.PwdHash.pbkdf2` now takes a `HashName` instead of a
+  `MacName`.
+* PATCH: update documentation in the `Botan.Low.PwdHash` module.
 
 ## 0.0.2.0 -- 2025-09-17